// src/modules/autenticacion/domain/aggregates/types.ts
export enum TipoUsuario {
  CLIENTE = 'CLIENTE',
  EMPLEADO = 'EMPLEADO',
}

export enum EstadoCuenta {
  ACTIVA = 'ACTIVA',
  INACTIVA = 'INACTIVA',
  BLOQUEADA = 'BLOQUEADA',
  PENDIENTE_VERIFICACION = 'PENDIENTE_VERIFICACION',
}

export enum EstadoSesion {
  ACTIVA = 'ACTIVA',
  REVOCADA = 'REVOCADA',
  EXPIRADA = 'EXPIRADA',
}

export enum TipoTokenRecuperacion {
  RECUPERACION_PASSWORD = 'RECUPERACION_PASSWORD',
  VERIFICACION_EMAIL = 'VERIFICACION_EMAIL',
}

export enum EstadoToken {
  PENDIENTE = 'PENDIENTE',
  USADO = 'USADO',
  EXPIRADO = 'EXPIRADO',
  INVALIDADO = 'INVALIDADO',
}

export enum TipoEventoAuth {
  LOGIN = 'LOGIN',
  LOGOUT = 'LOGOUT',
  REFRESH_TOKEN = 'REFRESH_TOKEN',
  CAMBIO_PASSWORD = 'CAMBIO_PASSWORD',
  RECUPERACION_PASSWORD_SOLICITUD = 'RECUPERACION_PASSWORD_SOLICITUD',
  RECUPERACION_PASSWORD_USO = 'RECUPERACION_PASSWORD_USO',
  VERIFICACION_EMAIL = 'VERIFICACION_EMAIL',
  BLOQUEO_CUENTA = 'BLOQUEO_CUENTA',
  DESBLOQUEO_CUENTA = 'DESBLOQUEO_CUENTA',
  REVOCACION_SESION = 'REVOCACION_SESION',
  REVOCACION_MASIVA = 'REVOCACION_MASIVA',
  CREACION_CUENTA_EMPLEADO = 'CREACION_CUENTA_EMPLEADO',
}

export enum ResultadoAuth {
  EXITOSO = 'EXITOSO',
  FALLIDO = 'FALLIDO',
}